Only load icon font once
The icon font is not exactly small, so this speeds up the build.

include_bytes! shouldn't be called twice on same data if it can at all
be helped.

Furthermore, putting it in a function also helps I've found, rather than
in the root where any module can ask for it.